Combine `xpath-util` and `xpath` into a single module
There was no good reason for the separation between the two any more.
